# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tontines": {
      "post": {
        "operationId": "TontinesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200
                  },
                  "type": {
                    "type": "string",
                    "enum": ["SIMPLE_ROTATIVE"],
                    "default": "SIMPLE_ROTATIVE"
                  },
                  "contributionAmount": {
                    "type": "string",
                    "pattern": "^(0|[1-9]\\d{0,14})(\\.\\d+)?$"
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  },
                  "frequency": {
                    "type": "string",
                    "enum": ["WEEKLY", "BIWEEKLY", "MONTHLY", "BIMONTHLY"]
                  },
                  "frequencyDetail": {
                    "type": "object",
                    "properties": {
                      "day": {
                        "type": "string",
                        "enum": [
                          "monday",
                          "tuesday",
                          "wednesday",
                          "thursday",
                          "friday",
                          "saturday",
                          "sunday"
                        ]
                      },
                      "weekOfMonth": {
                        "anyOf": [
                          {
                            "type": "number",
                            "enum": [1]
                          },
                          {
                            "type": "number",
                            "enum": [2]
                          },
                          {
                            "type": "number",
                            "enum": [3]
                          },
                          {
                            "type": "number",
                            "enum": [4]
                          },
                          {
                            "type": "number",
                            "enum": [-1]
                          }
                        ]
                      },
                      "lastDayOfMonth": {
                        "type": "boolean"
                      }
                    },
                    "additionalProperties": false,
                    "default": {}
                  },
                  "maxMembers": {
                    "type": "integer",
                    "minimum": 3,
                    "maximum": 50
                  },
                  "startDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "drawMode": {
                    "type": "string",
                    "enum": ["RANDOM", "FIXED_ORDER", "PRIORITY_NEED"]
                  },
                  "penaltyRules": {
                    "type": "object",
                    "properties": {
                      "graceDays": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 30
                      },
                      "lateFeePercent": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 100
                      },
                      "suspendAfter": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 12
                      },
                      "defaultAfterDays": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 60,
                        "default": 7
                      }
                    },
                    "required": ["graceDays", "lateFeePercent", "suspendAfter"],
                    "additionalProperties": false
                  },
                  "entryFee": {
                    "type": "string",
                    "pattern": "^(0|[1-9]\\d{0,14})(\\.\\d+)?$"
                  },
                  "collation": {
                    "type": "string",
                    "pattern": "^(0|[1-9]\\d{0,14})(\\.\\d+)?$"
                  },
                  "incompletePolicy": {
                    "type": "string",
                    "enum": ["POSTPONE", "PARTIAL_PAYOUT"],
                    "default": "POSTPONE"
                  }
                },
                "required": [
                  "name",
                  "contributionAmount",
                  "frequency",
                  "maxMembers",
                  "startDate",
                  "drawMode",
                  "penaltyRules"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une tontine (US-4.1) — KYC TIER_3 ou délégation du super-admin",
        "tags": ["Tontines"]
      },
      "get": {
        "operationId": "TontinesController_mine_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mes tontines (participant ou admin)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}": {
      "get": {
        "operationId": "TontinesController_one_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d’une tontine (participants et super-admin)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}/participants": {
      "get": {
        "operationId": "TontinesController_participants_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Participants (prénom, rôle, position — sans montants individuels)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}/cancel": {
      "post": {
        "operationId": "TontinesController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  }
                },
                "required": ["reason"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annuler une tontine avant démarrage (admin)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}/invitations": {
      "post": {
        "operationId": "TontinesController_invite_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "channel": {
                        "type": "string",
                        "enum": ["EMAIL"]
                      },
                      "email": {
                        "type": "string",
                        "maxLength": 255,
                        "format": "email"
                      }
                    },
                    "required": ["channel", "email"],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "channel": {
                        "type": "string",
                        "enum": ["PHONE"]
                      },
                      "phone": {
                        "type": "string"
                      }
                    },
                    "required": ["channel", "phone"],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "channel": {
                        "type": "string",
                        "enum": ["LINK"]
                      }
                    },
                    "required": ["channel"],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inviter par email, téléphone ou lien partageable (US-4.2)",
        "tags": ["Tontines"]
      },
      "get": {
        "operationId": "TontinesController_invitationsOf_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invitations de la tontine (admin)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}/invitations/{invitationId}": {
      "delete": {
        "operationId": "TontinesController_revoke_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Révoquer une invitation non acceptée (US-4.2 §7)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/me/invitations": {
      "get": {
        "operationId": "TontinesController_myInvitations_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invitations reçues en attente",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/invitations/{id}/respond": {
      "post": {
        "operationId": "TontinesController_respond_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accept": {
                    "type": "boolean"
                  }
                },
                "required": ["accept"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accepter ou refuser une invitation (éligibilité vérifiée à l’acceptation)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}/invitations/{invitationId}/accept": {
      "post": {
        "operationId": "TontinesController_accept_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accepter une invitation (alias de invitations/{id}/respond)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/tontines/{id}/invitations/{invitationId}/reject": {
      "post": {
        "operationId": "TontinesController_decline_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refuser une invitation (alias de invitations/{id}/respond)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/invitations/code/{code}": {
      "get": {
        "operationId": "TontinesController_preview_v1",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aperçu public d’une invitation (nom, montant, fréquence, date de début)",
        "tags": ["Tontines"]
      }
    },
    "/api/v1/invitations/code/{code}/accept": {
      "post": {
        "operationId": "TontinesController_acceptCode_v1",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rejoindre via un lien d’invitation",
        "tags": ["Tontines"]
      }
    },
...
  },
  "info": {
    "title": "TontineMoney API",
    "description": "API REST versionnée du MVP TontineMoney (monolithe modulaire). Erreurs au format Problem Details (RFC 9457). Montants en chaînes décimales (unités majeures) + devise ISO 4217. En-tête Idempotency-Key obligatoire sur les POST financiers.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tm_rt"
      }
    },
...
  }
}